Add monad trans control instances
