+++ gore-and-ash-logging.cabal
@@ -41,6 +41,7 @@ library
     , fast-logger           >= 2.4      && < 2.5
     , gore-and-ash          >= 2.0.0.0  && < 2.1
     , hashable              >= 1.2.4.0  && < 1.3
+    , monad-control         >= 1.0      && < 1.1
     , mtl                   >= 2.2      && < 2.3
     , reflex                >= 0.5      && < 0.6
     , resourcet             >= 1.1.7.1  && < 1.2
